Add auto-repair code if student in team is missing.
